Move ensure-array module to array subfolder
